+++ src/main/java/tools/jackson/databind/EnumNamingStrategy.java
@@ -13,9 +13,9 @@ public interface EnumNamingStrategy {
      * Translates the given <code>enumName</code> into an external property name according to
      * the implementation of this {@link EnumNamingStrategy}.
      *
-     * @param enumName the name of the enum value to translate
      * @param config the mapper configuration
      * @param cls the Enum class
+     * @param enumName the name of the enum value to translate
      *
      * @return the external property name that corresponds to the given <code>enumName</code>
      * according to the implementation of this {@link EnumNamingStrategy}.
